requirements.txt
streamlit>=1.24.0
transformers>=4.35.0
torch>=2.0.0
PyPDF2>=3.0.0
python-docx>=0.8.11
accelerate>=0.20.3
sentencepiece>=0.1.99
regex
pytz
tqdm